% tinyrox says don't edit this manually, but it can't stop you!
\name{add_one_control_brace}
\alias{add_one_control_brace}
\title{Add Braces to One Control Flow Statement}
\usage{
add_one_control_brace(code)
}
\arguments{
\item{code}{Code string.}
}
\value{
Modified code or NULL if no changes.
}
\description{
Add Braces to One Control Flow Statement
}
\keyword{internal}
