% tinyrox says don't edit this manually, but it can't stop you!
\name{display_width}
\alias{display_width}
\title{Get Tab-Expanded Line Length}
\usage{
display_width(line)
}
\arguments{
\item{line}{A single line of text.}
}
\value{
Display width of the line.
}
\description{
Returns the display width of a line, with tabs expanded to 8-column stops.
}
\keyword{internal}
