% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming.R
\name{fancy_names}
\alias{fancy_names}
\title{\code{fancy_names} returns a character vector with pretty names for tables.}
\usage{
fancy_names(x)
}
\arguments{
\item{A}{character vector. Normalized names.}
}
\value{
A character vector.
}
\description{
Changes what should be space to it (-, _, and .). First letter in word
to uppercase.
}
\examples{
fancy_names(c("a.good_name", "another-name"))
}

