% tinyrox says don't edit this manually, but it can't stop you!
\name{fix_else_placement}
\alias{fix_else_placement}
\title{Fix Else Placement}
\usage{
fix_else_placement(code)
}
\arguments{
\item{code}{Code string.}
}
\value{
Code with corrected else placement.
}
\description{
Ensures `else` appears on the same line as the closing brace.
}
\keyword{internal}
