% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{fnumber}
\alias{fnumber}
\title{Converts number or numeric to string or character with pretty format.}
\usage{
fnumber(number)
}
\arguments{
\item{number}{Numeric, integer or number.}
}
\value{
character vector with readable format for numbers.
}
\description{
Converts number or numeric to string or character with pretty format.
}
\examples{
fnumber(454667)
fnumber(456234.3434356)
fnumber(c(453535635, 364634.6245624, 43634.35235236, 32535234523))
}

