% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{fnumber.s}
\alias{fnumber.s}
\title{Converts number to string with pretty format.}
\usage{
fnumber.s(n)
}
\arguments{
\item{n}{Integer or float.}
}
\description{
Converts number to string with pretty format.
}
\examples{
fnumber.s(459877364)
fnumber.s(34556.45778)
}

