% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/format_parsed.R
\name{format_line}
\alias{format_line}
\title{Format a Single Line}
\usage{
format_line(line)
}
\arguments{
\item{line}{Character string of a single line.}
}
\value{
Formatted line.
}
\description{
Apply formatting rules to a single line of code.
}
\keyword{internal}
