% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/format_parsed.R
\name{format_spacing}
\alias{format_spacing}
\title{Format Spacing}
\usage{
format_spacing(code)
}
\arguments{
\item{code}{Code string.}
}
\value{
Code with corrected spacing.
}
\description{
Apply spacing rules to code.
}
\keyword{internal}
