% tinyrox says don't edit this manually, but it can't stop you!
\name{is_parseable_code}
\alias{is_parseable_code}
\title{Parse Gate for Transform Passes}
\usage{
is_parseable_code(code)
}
\arguments{
\item{code}{Code string.}
}
\value{
TRUE if code parses, FALSE otherwise.
}
\description{
Parse Gate for Transform Passes
}
\keyword{internal}
