% tinyrox says don't edit this manually, but it can't stop you!
\name{line_index_get}
\alias{line_index_get}
\title{Look Up Row Indices for a Line}
\usage{
line_index_get(lidx, line_num)
}
\arguments{
\item{lidx}{Line index from `build_line_index()`.}

\item{line_num}{Output line number.}
}
\value{
Integer vector of row indices, or `integer(0)` if none.
}
\description{
Look Up Row Indices for a Line
}
\keyword{internal}
