% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming.R
\name{normalize_names}
\alias{normalize_names}
\title{\code{normalize_names} returns a character vector with sql names conformant.}
\usage{
normalize_names(column_names)
}
\arguments{
\item{A}{character vector. Column names with invalid characters.}
}
\value{
A character vector.
}
\description{
Removes trailing spaces, changes multiple spaces for one, changes spaces and
"." for "_", transforms camelCase to camel_case, removes Spanish characters
-accents and ñ-, transforms all to lowercase.
}
\examples{
normalize_names(c("Hello  world   ", "another.name", "camelCase", "ALL"))
}

