% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{pnumber.s}
\alias{pnumber.s}
\title{Converts number to string with pretty format.}
\usage{
pnumber.s(n)
}
\arguments{
\item{n}{Integer or float.}
}
\description{
Converts number to string with pretty format.
}
\examples{
pnumber.s(459877364)
pnumber.s(34556.45778)
}

