% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{ptable}
\alias{ptable}
\title{Formats a data frame numeric and integer columns into a readable format.}
\usage{
ptable(df, fnames = TRUE)
}
\arguments{
\item{fnames}{By default, \code{fnames = TRUE} transforms the names of the
columns into readable names using fancy_names. To leave them identical use
\code{fnames = FALSE}.}

\item{data.frame}{df}
}
\description{
Formats a data frame numeric and integer columns into a readable format.
}
\examples{
options(scipen = 999, digits = 10) # disable scientific notation
df <- data.frame(
  big_numbers = rnorm(10, 1000000,1000000),
  big_integers = floor(rnorm(10, 1000000,1000000)),
  mixed.numbers = c(rnorm(5, 100, 50), rnorm(5, 100, 50))
)
df
ptable(df)
}

