% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/format_parsed.R
\name{spacing_before}
\alias{spacing_before}
\title{Determine Spacing Before Token}
\usage{
spacing_before(tok, prev)
}
\arguments{
\item{tok}{Current token row.}

\item{prev}{Previous token row or NULL.}
}
\value{
Logical, TRUE if space needed before token.
}
\description{
Determine Spacing Before Token
}
\keyword{internal}
