% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{total_percentage}
\alias{total_percentage}
\title{Get totals and percentages.}
\usage{
total_percentage(totals, percentages, ...)
}
\arguments{
\item{totals}{data.frame with totals.}

\item{percentages}{data.frame with percentages.}

\item{\dots}{Additional arguments.}
}
\description{
Get totals and percentages.
}
\examples{
library(tidyr)
totals <- data.frame(names = letters[1:5], x = rnorm(5, 500, 200), y = rnorm(5, 500, 200))
per <- tidyr::gather(totals, key = key, value = value, -names) \%>\% dplyr::group_by(key) \%>\% dplyr::mutate(percentage = value/sum(value) * 100) \%>\% dplyr::select(-value) \%>\% tidyr::spread(key = key, value = percentage) \%>\% data.frame()
total_percentage(totals, per)
total_percentage(totals, per, fnames = F)
}

