% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/format_parsed.R
\name{format_blank_lines}
\alias{format_blank_lines}
\title{Format Blank Lines}
\usage{
format_blank_lines(code)
}
\arguments{
\item{code}{Code string.}
}
\value{
Code with normalized blank lines.
}
\description{
Normalize blank lines between code blocks.
}
\keyword{internal}
