% tinyrox says don't edit this manually, but it can't stop you!
\name{format_line_tokens}
\alias{format_line_tokens}
\title{Format Tokens on a Single Line}
\usage{
format_line_tokens(tokens)
}
\arguments{
\item{tokens}{Data frame of tokens for one line.}
}
\value{
Formatted line content (no leading whitespace).
}
\description{
Format Tokens on a Single Line
}
\keyword{internal}
