% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/format_parsed.R
\name{reformat_function_defs}
\alias{reformat_function_defs}
\title{Reformat Function Definitions}
\usage{
reformat_function_defs(code)
}
\arguments{
\item{code}{Formatted code string.}
}
\value{
Code with reformatted function definitions.
}
\description{
Ensures function definitions follow style guide: - First argument on new
line after function( - Each argument on its own line - Closing ) \{ on
its own line
}
\keyword{internal}
