% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/format_parsed.R
\name{reformat_inline_if}
\alias{reformat_inline_if}
\title{Reformat Inline If-Else Statements}
\usage{
reformat_inline_if(code)
}
\arguments{
\item{code}{Formatted code string.}
}
\value{
Code with reformatted inline if-else.
}
\description{
Expands inline if-else to multi-line format per style guide: x <- if
(cond) a else b becomes: if (cond) \{ x <- a \} else \{ x <- b \}
}
\keyword{internal}
