% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/format_parsed.R
\name{reformat_one_function}
\alias{reformat_one_function}
\title{Reformat One Function Definition}
\usage{
reformat_one_function(code)
}
\arguments{
\item{code}{Code string.}
}
\value{
Modified code or NULL if no changes.
}
\description{
Reformat One Function Definition
}
\keyword{internal}
