% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/format_parsed.R
\name{reformat_one_inline_if}
\alias{reformat_one_inline_if}
\title{Reformat One Inline If-Else Statement}
\usage{
reformat_one_inline_if(code)
}
\arguments{
\item{code}{Code string.}
}
\value{
Modified code or NULL if no changes.
}
\description{
Reformat One Inline If-Else Statement
}
\keyword{internal}
