% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/rformat.R
\name{rformat}
\alias{rformat}
\title{Format R Code}
\usage{
rformat(code)
}
\arguments{
\item{code}{Character string of R code to format.}
}
\value{
Formatted code as a character string.
}
\description{
Format R code string according to base R style conventions.
}
\examples{
rformat("x<-1+2")
rformat("if(x>0){y<-1}")
}
